refactor(tests): rename _usage_fixtures.py to _client_fixtures.py
PR #154 generalized this module beyond usage-endpoint fixtures (it now also
builds ChatCompletion mocks for chat-completions tests) and updated the
docstring accordingly, but left the filename mismatched. Renaming to match.
